Move servers to bin/glusterd2/servers
