+++ bin/glusterd2/commands/peers/getpeers.go
@@ -3,8 +3,8 @@ package peercommands
 import (
 	"net/http"
 
+	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/peer"
-	restutils "github.com/gluster/glusterd2/servers/rest/utils"
 )
 
 func getPeersHandler(w http.ResponseWriter, r *http.Request) {
