+++ bin/glusterd2/commands/version/version.go
@@ -4,7 +4,7 @@ package versioncommands
 import (
 	"net/http"
 
-	restutils "github.com/gluster/glusterd2/servers/rest/utils"
+	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/version"
 )
