+++ bin/glusterd2/commands/volumes/commands.go
@@ -2,7 +2,7 @@
 package volumecommands
 
 import (
-	"github.com/gluster/glusterd2/servers/rest/route"
+	"github.com/gluster/glusterd2/bin/glusterd2/servers/rest/route"
 )
 
 // Command is a holding struct used to implement the GlusterD Command interface
