+++ bin/glusterd2/commands/volumes/common.go
@@ -5,10 +5,10 @@ import (
 	"strings"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/bin/glusterd2/servers/sunrpc"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
 	"github.com/gluster/glusterd2/pkg/utils"
-	"github.com/gluster/glusterd2/servers/sunrpc"
 	"github.com/gluster/glusterd2/volume"
 	"github.com/gluster/glusterd2/xlator"
