+++ bin/glusterd2/commands/volumes/volume-list.go
@@ -3,7 +3,7 @@ package volumecommands
 import (
 	"net/http"
 
-	restutils "github.com/gluster/glusterd2/servers/rest/utils"
+	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/volume"
 )
