+++ bin/glusterd2/servers/muxsrv/supervisor.go
@@ -1,8 +1,8 @@
 package muxsrv
 
 import (
-	"github.com/gluster/glusterd2/servers/rest"
-	"github.com/gluster/glusterd2/servers/sunrpc"
+	"github.com/gluster/glusterd2/bin/glusterd2/servers/rest"
+	"github.com/gluster/glusterd2/bin/glusterd2/servers/sunrpc"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/thejerf/suture"
