+++ plugins/hello/init.go
@@ -1,7 +1,7 @@
 package hello
 
 import (
-	"github.com/gluster/glusterd2/servers/rest/route"
+	"github.com/gluster/glusterd2/bin/glusterd2/servers/rest/route"
 	"github.com/prashanthpai/sunrpc"
 )
